+++ src/core/app.c
@@ -1,5 +1,6 @@
 #include "core/app.h"
 #include "core/errorHandler.h"
+#include "core/input.h"
 #include "glad/glad.h"
 #include <GLFW/glfw3.h>
 
@@ -8,6 +9,7 @@
 #include "renderer/functionRenderer.h"
 #include "renderer/graph.h"
 #include "textRenderer/text.h"
+#include "ui/inputField.h"
 #include "ui/ui.h"
 #include "ui/uiInternal.h"
 #include "utils/shaderUtils.h"
@@ -195,30 +197,62 @@ enum reh_error_code_e ra_AppRenderFrame(struct ra_app_context_t *ctx, struct rui
   if (err != ERR_SUCCESS) return err;
   err = rtr_RenderDebugInfo(ctx->textProgram, ctx->textVAO, ctx->textVBO, chars, 1.0f, textColor, functions);
 
-  rui_Begin(uiCtx);
+  if (isUiShown){
+    rui_Begin(uiCtx);
 
-  rui_RenderRect(uiCtx, 10, 10, "a", 150, 500, (struct rm_vec3_t){1.0f, 1.0f, 1.0f}, rui_RectOnHoverGreen, rui_RectOnClickBlue);
-  rui_RenderRect(uiCtx, 165, 10, "yo", 500, 100, (struct rm_vec3_t){1.0f, 1.0f, 1.0f}, rui_RectOnHoverRed, rui_RectOnClickGray);
-  rui_RenderRect(uiCtx, 70, 255, "square", 100, 100, (struct rm_vec3_t){0.0f, 0.0f, 1.0f}, rui_RectOnHoverGray, rui_RectOnClickGreen);
+    rui_RenderRect(uiCtx, 10, 10, "a", 150, 500, (struct rm_vec3_t){1.0f, 1.0f, 1.0f}, rui_RectOnHoverGreen, rui_RectOnClickBlue);
+    rui_RenderRect(uiCtx, 165, 10, "yo", 500, 100, (struct rm_vec3_t){1.0f, 1.0f, 1.0f}, rui_RectOnHoverRed, rui_RectOnClickGray);
+    rui_RenderRect(uiCtx, 70, 255, "square", 100, 100, (struct rm_vec3_t){0.0f, 0.0f, 1.0f}, rui_RectOnHoverGray, rui_RectOnClickGreen);
 
-  int id;
-  rui_AACursorCheck(&id);
+    rui_RenderInputField(uiCtx, 100, 300, "f input", 300, 100, 5, 5, (struct rm_vec3_t){0.0f, 0.0f, 1.0f},(struct rm_vec3_t){1.0f, 0.0f, 0.0f}, rui_InputFieldOnClick);
+    rui_RenderInputField(uiCtx, 100, 100, "l input", 30, 10, 5, 5, (struct rm_vec3_t){0.0f, 0.0f, 1.0f},(struct rm_vec3_t){1.0f, 0.0f, 0.0f}, rui_InputFieldOnClick);
 
-  if (ruiState.mouseDown){
-    ruiState.activeItem = id;
-    for (size_t i = 0; i < uiCtx->commandCount; i++){
-      struct rui_command_t *currentCommand = &uiCtx->commands[i];
+    int id;
+    rui_AACursorCheck(&id);
 
-      if (ruiState.activeItem == currentCommand->id) currentCommand->rect.onClick(&currentCommand->rect);
+    // i ought to optimize this (TODO)
+    if (g_inputFieldData.isAnyPressed == true){
+      for (size_t i = 0; i < uiCtx->commandCount; i++){
+        struct rui_command_t *currentCommand = &uiCtx->commands[i];
+        if (currentCommand->type == RUI_INPUT_FIELD && currentCommand->id == g_inputFieldData.activeId){
+          currentCommand->input.onClick(&currentCommand->input);
+        }
+      }
     }
-  }
-  else {
-    ruiState.hotItem = id;
 
-    for (size_t i = 0; i < uiCtx->commandCount; i++){
-      struct rui_command_t *currentCommand = &uiCtx->commands[i];
+    if (ruiState.mouseDown){
+      ruiState.activeItem = id;
+      for (size_t i = 0; i < uiCtx->commandCount; i++){
+        struct rui_command_t *currentCommand = &uiCtx->commands[i];
 
-      if (ruiState.hotItem == currentCommand->id) currentCommand->rect.onHover(&currentCommand->rect);
+        switch (currentCommand->type){
+          case RUI_RECT:
+            if (ruiState.activeItem == currentCommand->id) currentCommand->rect.onClick(&currentCommand->rect);
+            break;
+          case RUI_INPUT_FIELD:
+            if (ruiState.activeItem == currentCommand->id){
+              currentCommand->input.onClick(&currentCommand->input);
+              g_inputFieldData.activeId = currentCommand->id;
+            }
+          default:
+            break;
+        }
+      }
+    }
+    else {
+      ruiState.hotItem = id;
+
+      for (size_t i = 0; i < uiCtx->commandCount; i++){
+        struct rui_command_t *currentCommand = &uiCtx->commands[i];
+
+        switch (currentCommand->type){
+          case RUI_RECT: 
+            if (ruiState.hotItem == currentCommand->id) currentCommand->rect.onHover(&currentCommand->rect);
+            break;
+          default:
+            break;
+        }
+      }
     }
   }
 
@@ -229,8 +263,7 @@ enum reh_error_code_e ra_AppRenderFrame(struct ra_app_context_t *ctx, struct rui
   rm_Mat4ValuePtr(&textProjection, &textProjectionPtr);
   rsu_GluSetMat4(ctx->textProgram, "textProjection", textProjectionPtr);
 
-
-  rui_End(uiCtx, &ctx->uiProgram, &ctx->uiVAO, &ctx->uiVBO, &ctx->uiEBO, &textProjectionPtr);
+  if (isUiShown) rui_End(uiCtx, &ctx->uiProgram, &ctx->uiVAO, &ctx->uiVBO, &ctx->uiEBO, &textProjectionPtr);
 
   return ERR_SUCCESS;
 }